requirements.txt
tqdm
PyYAML
numpy
pandas
matplotlib
seaborn
mlflow
pillow
scikit_learn
torch
captum
evaluate
facexlib
opencv_python
cmake
dlib
einops
transformers
mediapipe
imutils
scipy
torchvision==0.16.0
torchcam
